move methods to avoid import cycle
